everything include and build

+++ src/units/units.hpp
@@ -7,42 +7,42 @@
 #include "ratio.hpp"
 
 namespace math {
-    using Milimeters = types::MilimetersT<double>;
-    using Pixels = types::PixelsT<double>;
+    using Milimeters = types::MilimetersT<float>;
+    using Pixels = types::PixelsT<float>;
 
     namespace units {
         /* Pixels */
         constexpr Pixels operator "" _px(long double d) {
-            return Pixels{static_cast<double>(d)};
+            return Pixels{static_cast<float>(d)};
         }
 
         constexpr Pixels operator "" _px(unsigned long long d) {
-            return Pixels{static_cast<double>(d)};
+            return Pixels{static_cast<float>(d)};
         }
 
         /* Milimeters */
         constexpr Milimeters operator "" _mm(long double d) {
-            return Milimeters{static_cast<double>(d)};
+            return Milimeters{static_cast<float>(d)};
         }
 
         constexpr Milimeters operator "" _mm(unsigned long long d) {
-            return Milimeters{static_cast<double>(d)};
+            return Milimeters{static_cast<float>(d)};
         }
 
         constexpr Milimeters operator "" _cm(long double d) {
-            return Milimeters{static_cast<double>(d * 10)};
+            return Milimeters{static_cast<float>(d * 10)};
         }
 
         constexpr Milimeters operator "" _cm(unsigned long long d) {
-            return Milimeters{static_cast<double>(d * 10)};
+            return Milimeters{static_cast<float>(d * 10)};
         }
 
         constexpr Milimeters operator "" _m(long double d) {
-            return Milimeters{static_cast<double>(d * 1000)};
+            return Milimeters{static_cast<float>(d * 1000)};
         }
 
         constexpr Milimeters operator "" _m(unsigned long long d) {
-            return Milimeters{static_cast<double>(d * 1000)};
+            return Milimeters{static_cast<float>(d * 1000)};
         }
 
         // Degrees
